Update DB from float to decimal Add UnitTest

+++ WebKalkulacka/MathCalculation/Models/CalculationModel.cs
@@ -8,10 +8,10 @@ namespace MathCalculation.Models
 {
     public class CalculationModel
     {
-        public double Number1 { get; set; }
-        public double Number2 { get; set; }
+        public decimal Number1 { get; set; }
+        public decimal Number2 { get; set; }
         public Operation? Operation { get; set; }
-        public double Result { get; set; }
+        public decimal Result { get; set; }
     }
 
     public enum Operation
